Time: 33 ms (29.15%), Space: 42.9 MB (28.59%) - LeetHub

+++ 0115-distinct-subsequences/0115-distinct-subsequences.cpp
@@ -6,13 +6,9 @@ public:
         
         const int MOD = 1e9 + 7;
         vector<vector<long long>> dp(n+1, vector<long long>(m+1, 0));
-        
-        // Base case: An empty string `t` is a subsequence of any prefix of `s`
         for(int j = 0; j <= m; j++) {
             dp[0][j] = 1;
         }
-        
-        // Fill the dp table
         for(int i = 1; i <= n; i++) {
             for(int j = 1; j <= m; j++) {
                 if(s[j-1] == t[i-1]) {
@@ -22,8 +18,6 @@ public:
                 }
             }
         }
-
-        // Result is the number of distinct subsequences of `t` in `s`
-        return static_cast<int>(dp[n][m]);
+        return dp[n][m];
     }
 };
